# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIKMFKG YARSI",
        "description": "Sistem Informasi Kepaniteraan Mahasiswa",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v2/master/hospitals/select": {
            "get": {
                "tags": [
                    "Master Hospital"
                ],
                "summary": "Select",
                "operationId": "227a5cf0f2264e24bd66959f1630e4f0",
                "responses": {
                    "200": {
                        "description": "Success."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/master/hospitals": {
            "get": {
                "tags": [
                    "Master Hospital"
                ],
                "summary": "Index",
                "operationId": "402e0624d902a99dd1e5efe5a41af399",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Hospital's Pagination Limit.",
                        "required": false
                    },
                    {
                        "name": "current_page",
                        "in": "query",
                        "description": "Hospital's Pagination Current Page.",
                        "required": false
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Hospital's Pagination Order.",
                        "required": false
                    },
                    {
                        "name": "filter[]",
                        "in": "query",
                        "description": "Hospital's Pagination Filter.",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Master Hospital"
                ],
                "summary": "Store",
                "operationId": "19b0afaa6a32c6f1f93c40432e94cc7c",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Hospital's Name.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/master/hospitals/{id}": {
            "get": {
                "tags": [
                    "Master Hospital"
                ],
                "summary": "Show",
                "operationId": "a17c82936336224fedac7b324d2abba5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Hospital's Identifier.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success."
                    },
                    "404": {
                        "description": "Not Found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Master Hospital"
                ],
                "summary": "Update",
                "operationId": "91c915bd59409f246ce67a3f625663cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Hospital's Identifier.",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Hospital's Name.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "404": {
                        "description": "Not Found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Master Hospital"
                ],
                "summary": "Destroy",
                "operationId": "9421397a938068270fdce2c63f689b31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Hospital's Identifier.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "404": {
                        "description": "Not Found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/master/hospitals/activate/{id}": {
            "put": {
                "tags": [
                    "Master Hospital"
                ],
                "summary": "Activate",
                "operationId": "ac1882e609ed95b16b7b267b771c1823",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Hospital's Identifier.",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "404": {
                        "description": "Not Found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/master/hospitals/deactivate/{id}": {
            "put": {
                "tags": [
                    "Master Hospital"
                ],
                "summary": "Deactivate",
                "operationId": "2c7468e54fdaa238c621b30bfee06be6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Hospital's Identifier.",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "404": {
                        "description": "Not Found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}